move utils not directly related to engines back to v1/utils.py
